requirements.txt
# Core dependencies
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-text-splitters==0.3.4

# Vector database
chromadb==0.5.23

# Local embedding model
sentence-transformers==3.3.1

# Environment management
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
